debugger: initial implementation of peek/poke
We can now communicate with the card over long periods of time.
That is, we now have an interactive shell where we can peek/poke
memory.  This commit doesn't have "jump" capabilities yet, but that
will come soon.

It appears to remain stable even after issuing many different commands,
which is very promising for creating a more interactive debugger.