main: ignore response from card during boot
We used to expect a specific response from the card.  This response
seems to vary from board to board.  Since we do a search of RAM for
magic values on boot anyway, we can very easily determine if we have
successfully entered debug mode without looking for a specific initial
response.